Add constants in Routes, fixed js scripts and add disable until fetch request, refactor all controllers for new routes

+++ src/main/java/com/case_opener_game/case_opener/dto/bet/BetResponseDTO.java
@@ -12,6 +12,6 @@ import java.math.BigDecimal;
 public class BetResponseDTO {
     private Long id;
     private BigDecimal amount;
-    private int multiplier;
+    private BigDecimal multiplier;
     private BigDecimal payout;
 }
