+++ src/main/java/com/case_opener_game/case_opener/entity/Round.java
@@ -26,7 +26,7 @@ public class Round {
     private BigDecimal amount;
 
     @Column(name = "multiplier")
-    private Integer multiplier;
+    private BigDecimal multiplier;
 
     @Column(name = "payout")
     private BigDecimal payout;
